// frontend/src/app/core/api/model/employeeResponse.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EmployeeResponseDto { 
    id?: string;
    firstName?: string;
    lastName?: string;
    fullName?: string;
    phoneNumber?: string;
    email?: string;
    salaryAmount?: number;
    salaryCurrency?: string;
    role?: string;
    employmentStatus?: EmployeeResponseDto.EmploymentStatusEnum;
    firstWorkingDay?: string;
    emergencyContactName?: string;
    emergencyContactPhone?: string;
    notes?: string;
    lastPaymentDate?: string;
    payrollRecordsCount?: number;
    monthsEmployed?: number;
}
export namespace EmployeeResponseDto {
    export const EmploymentStatusEnum = {
        Active: 'ACTIVE',
        Inactive: 'INACTIVE'
    } as const;
    export type EmploymentStatusEnum = typeof EmploymentStatusEnum[keyof typeof EmploymentStatusEnum];
}